Move notificationService in Mailer folder
